+++ src/scopes.rs
@@ -55,7 +55,17 @@ macro_rules! scope_impls {
             #[doc = "Please note that this may not work for you, as some auth flows and \"apis\" don't accept all scopes"]
             pub fn all() -> Vec<Scope> {
                 vec![
+                    $(
+                        scope_impls!(@omit $(#[deprecated($depr)])* $i),
+                    )*
+                ]
+            }
 
+            #[doc = "Get a slice of all defined twitch [Scopes][Scope]."]
+            #[doc = "\n\n"]
+            #[doc = "Please note that this may not work for you, as some auth flows and \"apis\" don't accept all scopes"]
+            pub const fn all_slice() -> &'static [Scope] {
+                &[
                     $(
                         scope_impls!(@omit $(#[deprecated($depr)])* $i),
                     )*
@@ -63,7 +73,7 @@ macro_rules! scope_impls {
             }
 
             #[doc = "Get a description for the token"]
-            pub fn description(&self) -> &'static str {
+            pub const fn description(&self) -> &'static str {
                 #![allow(deprecated)]
 
                 match self {
@@ -99,11 +109,44 @@ macro_rules! scope_impls {
                         $(#[cfg($cfg)])*
                         Scope::$i => $rename,
                     )*
-                    Self::Other(c) =>  c.as_ref()
+                    Self::Other(c) => c.as_ref()
+                }
+            }
+
+            /// Get the scope as a static string slice.
+            ///
+            /// ## Panics
+            ///
+            /// This function panics if the scope can't be represented as a static string slice
+            pub const fn as_static_str(&self) -> &'static str {
+                #![allow(deprecated)]
+                match self {
+                    $(
+                        $(#[cfg($cfg)])*
+                        Scope::$i => $rename,
+                    )*
+                    Self::Other(Cow::Borrowed(s)) => s,
+                    _ => panic!(),
                 }
             }
         }
-
+        #[test]
+        #[cfg(test)]
+        fn sorted() {
+            let slice = [$(
+                $(#[cfg($cfg)])*
+                $rename,
+            )*];
+            let mut slice_sorted = [$(
+                $(#[cfg($cfg)])*
+                $rename,
+            )*];
+            slice_sorted.sort();
+            for (scope, sorted) in slice.iter().zip(slice_sorted.iter()) {
+                assert_eq!(scope, sorted);
+            }
+            assert_eq!(slice, slice_sorted);
+        }
     };
 }
 
@@ -146,7 +189,7 @@ scope_impls!(
     ModeratorManageBlockedTerms,    scope: "moderator:manage:blocked_terms",    doc: "Manage a broadcaster’s list of blocked terms.";
     ModeratorManageChatMessages,    scope: "moderator:manage:chat_messages",    doc: "Delete chat messages in channels where you have the moderator role";
     ModeratorManageChatSettings,    scope: "moderator:manage:chat_settings",    doc: "View a broadcaster’s chat room settings.";
-    ModeratorReadShoutouts,         scope: "moderator:read:shoutouts",          doc: "View a broadcaster’s shoutouts.";
+    ModeratorManageShieldMode,      scope: "moderator:manage:shield_mode",      doc: "Manage a broadcaster’s Shield Mode status.";
     ModeratorManageShoutouts,       scope: "moderator:manage:shoutouts",        doc: "Manage a broadcaster’s shoutouts.";
     ModeratorReadAutomodSettings,   scope: "moderator:read:automod_settings",   doc: "View a broadcaster’s AutoMod settings.";
     ModeratorReadBlockedTerms,      scope: "moderator:read:blocked_terms",      doc: "View a broadcaster’s list of blocked terms.";
@@ -154,7 +197,7 @@ scope_impls!(
     ModeratorReadChatters,          scope: "moderator:read:chatters",           doc: "View the chatters in a broadcaster’s chat room.";
     ModeratorReadFollowers,         scope: "moderator:read:followers",          doc: "Read the followers of a broadcaster.";
     ModeratorReadShieldMode,        scope: "moderator:read:shield_mode",        doc: "View a broadcaster’s Shield Mode status.";
-    ModeratorManageShieldMode,      scope: "moderator:manage:shield_mode",      doc: "Manage a broadcaster’s Shield Mode status.";
+    ModeratorReadShoutouts,         scope: "moderator:read:shoutouts",          doc: "View a broadcaster’s shoutouts.";
     UserEdit,                       scope: "user:edit",                         doc: "Manage a user object.";
     UserEditBroadcast,              scope: "user:edit:broadcast",               doc: "Edit your channel's broadcast configuration, including extension configuration. (This scope implies user:read:broadcast capability.)";
     #[deprecated(note = "Not used anymore, see https://discuss.dev.twitch.tv/t/deprecation-of-create-and-delete-follows-api-endpoints/32351")]
@@ -205,7 +248,7 @@ mod tests {
     #[allow(deprecated)]
     fn no_deprecated() {
         for scope in Scope::all() {
-            assert!(scope != Scope::ChannelSubscriptions )
+            assert!(scope != Scope::ChannelSubscriptions)
         }
     }
 }